requirements.txt
absl-py==2.3.0
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aiosignal==1.4.0
alembic==1.16.4
altair==5.5.0
amqp==5.3.1
annotated-types==0.7.0
ansicolors==1.1.8
antlr4-python3-runtime==4.9.3
anyio==4.9.0
appdirs==1.4.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arro3-core==0.6.3
arrow==1.3.0
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.5
async-timeout==5.0.1
asyncssh==2.21.0
atpublic==6.0.1
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.5
billiard==4.2.1
bleach==6.2.0
blinker==1.9.0
blis==0.7.11
boto3==1.40.30
botocore==1.40.30
cachetools==5.5.2
catalogue==2.0.10
celery==5.5.3
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudpathlib==0.16.0
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
confection==0.1.5
configobj==5.0.9
contourpy==1.3.2
cramjam==2.10.0
cryptography==45.0.5
cycler==0.12.1
cymem==2.0.11
databricks-sdk==0.61.0
datasets==4.0.0
dbfread==2.0.7
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
delta-spark==3.1.0
deltalake==1.1.4
Deprecated==1.2.18
dictdiffer==0.9.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
dpath==2.2.0
duckdb==1.4.0
dulwich==0.24.1
dvc==3.61.0
dvc-data==3.16.10
dvc-http==2.32.0
dvc-objects==5.1.1
dvc-render==1.0.2
dvc-studio-client==0.22.0
dvc-task==0.40.2
elastic-transport==8.17.1
elasticsearch==9.0.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
entrypoints==0.4
evaluate==0.4.5
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.116.1
fastjsonschema==2.21.1
fastparquet==2024.11.0
fief-client==0.20.0
filelock==3.18.0
findspark==2.0.1
Flask==3.1.1
flatbuffers==25.2.10
flatten-dict==0.4.2
flufl.lock==8.2.0
fonttools==4.58.2
fqdn==1.5.1
frozenlist==1.7.0
fsspec==2025.3.0
funcy==2.0
gast==0.6.0
gensim==4.3.3
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.40.3
google-pasta==0.2.0
grandalf==0.8
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.2.3
grpcio==1.72.1
gto==1.7.2
gunicorn==23.0.0
h11==0.16.0
h5py==3.14.0
halo==0.0.31
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.27.2
huggingface-hub==0.34.4
hydra-core==1.3.2
idna==3.10
importlib_metadata==8.7.0
instaloader==4.14.1
interchange==2021.0.4
ipykernel==6.30.1
ipython==8.37.0
ipywidgets==8.1.7
isoduration==20.11.0
iterative-telemetry==0.0.10
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
jwcrypto==1.5.6
kaggle==1.7.4.5
keras==3.10.0
kiwisolver==1.4.8
kombu==5.5.4
langcodes==3.5.0
langdetect==1.0.9
language_data==1.3.0
libclang==18.1.1
log-symbols==0.0.14
Mako==1.3.10
marisa-trie==1.3.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.4
ml_dtypes==0.5.1
mlflow==3.1.4
mlflow-skinny==3.1.4
monotonic==1.6
mpmath==1.3.0
multidict==6.6.3
multiprocess==0.70.16
murmurhash==1.0.13
namex==0.1.0
narwhals==1.44.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
notebook==7.4.5
notebook_shim==0.2.4
numexpr==2.10.2
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-cusparselt-cu12==0.7.1
nvidia-ml-py==12.575.51
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
opentelemetry-api==1.36.0
opentelemetry-sdk==1.36.0
opentelemetry-semantic-conventions==0.57b0
opt_einsum==3.4.0
optree==0.16.0
orjson==3.11.1
overrides==7.7.0
packaging==25.0
pandas==2.3.1
pandocfilters==1.5.1
pansi==2024.11.0
papermill==2.6.0
parquet_tools==0.2.16
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
preshed==3.0.10
prometheus_client==0.22.1
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==5.29.5
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
py2neo==2021.2.4
py4j==0.10.9.7
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydot==4.0.1
pygit2==1.18.1
Pygments==2.19.1
pygtrie==2.5.0
pyLDAvis==3.4.1
pynvml==12.0.0
pyparsing==3.2.3
PySocks==1.7.1
pyspark==3.5.3
python-dateutil==2.9.0.post0
python-json-logger==3.3.0
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.4.0
questionary==2.1.0
RapidFuzz==3.13.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rpds-py==0.25.1
rpy2==3.6.0
rpy2-rinterface==3.6.0
rpy2-robjects==3.6.0
rsa==4.9.1
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
s3transfer==0.14.0
safetensors==0.5.3
scikit-learn==1.3.2
scipy==1.10.1
scmrepo==3.5.1
seaborn==0.13.2
semver==3.0.4
Send2Trash==1.8.3
sentence-transformers==2.7.0
sentencepiece==0.2.1
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.2
six==1.17.0
smart-open==6.4.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spinners==0.0.24
SQLAlchemy==2.0.42
sqlparse==0.5.3
sqltrie==0.11.2
srsly==2.5.1
stack-data==0.6.3
starlette==0.47.2
sympy==1.13.1
tabulate==0.9.0
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.3.0
terminado==0.18.1
text-unidecode==1.3
textblob==0.19.0
tf_keras==2.19.0
thinc==8.2.5
threadpoolctl==3.6.0
thrift==0.16.0
tinycss2==1.4.0
tokenizers==0.22.0
tomli==2.2.1
tomlkit==0.13.3
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.56.0
triton==3.1.0
typer==0.9.4
types-python-dateutil==2.9.0.20250809
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
uri-template==1.3.0
urllib3==2.4.0
uvicorn==0.35.0
vega-datasets==0.9.0
vine==5.1.0
voluptuous==0.15.2
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.3.4
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
widgetsnbextension==4.0.14
wordcloud==1.9.4
wrapt==1.17.3
xxhash==3.5.0
yarl==1.20.1
yaspin==3.1.0
zc.lockfile==3.0.post1
zipp==3.22.0
